Add language files to repo (#594)
* add language files and cleanup package.json

* hopefully this allows QA to proceed?

+++ source/SIL.AppBuilder.Portal.Frontend/scripts/refresh-languages.sh
@@ -12,7 +12,6 @@
 # NOTE: the CWD/PWD should be the frontend folder. not scripts.
 set -e
 
-
 url="https://raw.githubusercontent.com/silnrsi/sldr/master/extras/langtags.json"
 folder="./src/public/assets/language/"
 tempPath="${folder}langtags.tmp.json"
@@ -21,12 +20,14 @@ target2Path="${folder}specialtags.json"
 
 mkdir -p $folder
 
-echo "downloading all supported languages..."
-wget --output-document $tempPath $url
-node scripts/split-langtags-json.js $tempPath $target1Path $target2Path
-echo "downloaded all supported languages"
-
-echo "downloading transaltion support for the language data..."
+if [ ! -f $target1Path ]; then
+  echo "downloading all supported languages..."
+  wget --output-document $tempPath $url
+  node scripts/split-langtags-json.js $tempPath $target1Path $target2Path
+  echo "downloaded all supported languages"
+else 
+  echo "langtags.json has already been downloaded and processed"
+fi
 
 function downloadAndConvert {
   local lang=$1
@@ -34,16 +35,24 @@ function downloadAndConvert {
   local tmpName="${folder}ldml.tmp.$fileName.json"
   local finalName="${folder}${fileName}/ldml.json"
 
-  mkdir -p "${folder}${fileName}"
 
-  wget -qO- "https://ldml.api.sil.org/$lang?inc[0]=localeDisplayNames" | npx fast-xml-parser -o $tmpName
+  if [ ! -f $finalName ]; then
+    echo "downloading ldml data for $lang..."
 
-  # TODO: use a custom node script to convert the language
-  # list to something more easily consumeable by a javascript app.
-  node scripts/clean-up-ldml-json.js $tmpName $finalName
+    mkdir -p "${folder}${fileName}"
 
-  rm $tmpName
+    wget -qO- "https://ldml.api.sil.org/$lang?inc[0]=localeDisplayNames" | npx fast-xml-parser -o $tmpName
+
+    # TODO: use a custom node script to convert the language
+    # list to something more easily consumeable by a javascript app.
+    node scripts/clean-up-ldml-json.js $tmpName $finalName
+
+    rm $tmpName
+  else
+    echo "$fileName has already been downloaded and processed"
+  fi
 }
+
 
 downloadAndConvert "es-419"
 # TODO: should we rename our en-us translations to en?
